test: move soft-delete lookup tests into MemberPersistenceServiceIT
