+++ backend/src/test/java/ch/puzzle/pcts/service/persistence/MemberPersistenceServiceIT.java
@@ -82,4 +82,36 @@ class MemberPersistenceServiceIT extends PersistenceBaseIT<Member, MemberReposit
 
         assertTrue(result.isEmpty());
     }
+
+    @DisplayName("Should not return member by id when member is deleted")
+    @Test
+    void shouldNotReturnMemberByIdWhenIsDeleted() {
+        Optional<Member> result = persistenceService.findById(DELETED_MEMBER_4.getId());
+
+        assertTrue(result.isEmpty());
+    }
+
+    @DisplayName("Should not return member by LDAP name when member is deleted")
+    @Test
+    void shouldNotReturnMemberByLdapNameWhenIsDeleted() {
+        Optional<Member> result = persistenceService.findByLdapName(DELETED_MEMBER_4.getLdapName());
+
+        assertTrue(result.isEmpty());
+    }
+
+    @DisplayName("Should not return member by ptimeId when member is deleted")
+    @Test
+    void shouldNotReturnMemberByPtimeIdWhenIsDeleted() {
+        Optional<Member> result = persistenceService.findByPtimeId(DELETED_MEMBER_4.getPtimeId());
+
+        assertTrue(result.isEmpty());
+    }
+
+    @DisplayName("Should not return member by abbreviation when member is deleted")
+    @Test
+    void shouldNotReturnMemberByAbbreviationWhenIsDeleted() {
+        Optional<Member> result = persistenceService.findByAbbreviation(DELETED_MEMBER_4.getAbbreviation());
+
+        assertTrue(result.isEmpty());
+    }
 }
